python/requirements-dev.txt
azure-storage-blob
hypothesis
msal
numpy
pytest
pytest-runner
requests
requests_mock
segyio
tqdm
ujson
